Upgrade to newer protocol 2.3b

+++ Drivers/Nano-OpenTherm/Responses/DhwSetpointBoundsResponse.cs
@@ -10,18 +10,18 @@ namespace TekuSP.Drivers.Nano_OpenTherm.Responses
         public DhwSetpointBoundsResponse(MessageType mt = MessageType.READ_ACK) { MessageType = mt; }
         public DhwSetpointBoundsResponse(Response r) : base(r) { }
 
-        public byte UpperBoundC { get; set; }
-        public byte LowerBoundC { get; set; }
+    public sbyte UpperBoundC { get; set; }
+    public sbyte LowerBoundC { get; set; }
 
         protected override uint GetRawDataCore()
         {
-            ushort payload = Utilities.MakeUShort(UpperBoundC, LowerBoundC);
+            ushort payload = Utilities.MakeUShort((byte)UpperBoundC, (byte)LowerBoundC);
             return ProcessResponse(payload);
         }
         protected override void SetRawDataCore(uint value)
         {
-            UpperBoundC = Utilities.GetHighByte(value);
-            LowerBoundC = Utilities.GetLowByte(value);
+            UpperBoundC = (sbyte)Utilities.GetHighByte(value);
+            LowerBoundC = (sbyte)Utilities.GetLowByte(value);
         }
         public override MessageType MessageType { get; set; }
         public override MessageID MessageID => MessageID.TdhwSetUBTdhwSetLB;
